pyproject.toml
[build-system]
requires = ["setuptools>=70.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "expert-graph-rag"
version = "0.1.0"
description = "Production-grade demo: Django + pgvector + Neo4j + Celery"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "Django>=5.1,<6.0",
  "djangorestframework>=3.15,<4.0",
  "gunicorn>=23.0,<24.0",
  "psycopg[binary]>=3.1,<4.0",
  "celery[redis]>=5.4,<6.0",
  "redis>=5.0,<6.0",
  "neo4j>=5.23,<6.0",
  "pgvector>=0.3.6,<0.4",
  "sentence-transformers>=3.0,<4.0",
  "openai>=1.54,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "pytest-django>=4.9,<5.0",
  "ruff>=0.8,<1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["config*", "apps*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["test_*.py", "*_tests.py"]
markers = [
  "integration: marks tests that require Docker Compose and external services",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "DJ"]
